Fix reset of repository after conflicts.
There was an issue after a merge conflict: When the conflicting files
were not unstaged in the local repository, jgit refuses to checkout
master, even using the force flag.

After inspecting the source code of jgit, I think that this is really
they way they want jgit to behave, as they explicitly call
setFailOnConflict(true) without giving us a chance to override this
setting.

As there is nothing like `git merge --abort` in jgit, we have to deal
with this using the reset command. There is a bug related to this issue
in the jgit bugtracker, where OP seems to have solved this issue:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=428303

The only difference to our code was that they explicitly set the
reference to origin/master, which indeed seems to do the trick.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>